Add a module for style functions.
